Add utility function for migrating icon name based of outputFormat
